move metadata creation and modification into local queues, crud_metadata_ingest contains functions for metadata
